Refactor HouseKeeperHelper
