+++ TestChallengeProjects.sln
@@ -3,9 +3,12 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.32112.339
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestChallengeProjects", "TestChallengeProjects\TestChallengeProjects.csproj", "{F308DF9D-B016-4C30-9B81-8375036F04B3}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestChallengeProjects", "TestChallengeProjects\TestChallengeProjects.csproj", "{F308DF9D-B016-4C30-9B81-8375036F04B3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestChallengeProjectsTests", "TestChallengeProjectsTests\TestChallengeProjectsTests.csproj", "{91C1B9FD-2367-4449-BA65-EE79E54DA695}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestChallengeProjectsTests", "TestChallengeProjectsTests\TestChallengeProjectsTests.csproj", "{91C1B9FD-2367-4449-BA65-EE79E54DA695}"
+	ProjectSection(ProjectDependencies) = postProject
+		{F308DF9D-B016-4C30-9B81-8375036F04B3} = {F308DF9D-B016-4C30-9B81-8375036F04B3}
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
