+++ TestChallengeProjects/Common/IUnitOfWork.cs
@@ -1,0 +1,8 @@
+﻿
+namespace TestChallengeProjects
+{
+    public interface IUnitOfWork
+    {
+        IQueryable<T> Query<T>();
+    }
+}
