+++ TestChallengeProjects/Common/UnitOfWork.cs
@@ -1,6 +1,6 @@
 ﻿namespace TestChallengeProjects
 {
-    public class UnitOfWork
+    public class UnitOfWork : IUnitOfWork
     {
         public IQueryable<T> Query<T>()
         {
