--- TestChallengeProjects/HouseKeeperHelperProject/HouseKeeperHelper.cs
@@ -1,91 +1,0 @@
-﻿using System.Net;
-using System.Net.Mail;
-using System.Text;
-
-namespace TestChallengeProjects
-{
-    public static class HouseKeeperHelper
-    {
-        private static readonly UnitOfWork UnitOfWork = new UnitOfWork();
-
-        public static bool SendStatementEmails(DateTime statementDate)
-        {
-            var housekeepers = UnitOfWork.Query<Housekeeper>();
-
-            foreach (var housekeeper in housekeepers)
-            {
-                if (housekeeper.Email == null)
-                    continue;
-
-                var statementFilename = SaveStatement(housekeeper.Oid, housekeeper.FullName, statementDate);
-
-                if (string.IsNullOrWhiteSpace(statementFilename))
-                    continue;
-
-                var emailAddress = housekeeper.Email;
-                var emailBody = housekeeper.StatementEmailBody;
-
-                try
-                {
-                    EmailFile(emailAddress, emailBody, statementFilename,
-                        string.Format("Sandpiper Statement {0:yyyy-MM} {1}", statementDate, housekeeper.FullName));
-                }
-                catch (Exception e)
-                {
-                    XtraMessageBox.Show(e.Message, string.Format("Email failure: {0}", emailAddress),
-                        MessageBoxButtons.OK);
-                }
-            }
-
-            return true;
-        }
-
-        private static string SaveStatement(int housekeeperOid, string housekeeperName, DateTime statementDate)
-        {
-            var report = new HousekeeperStatementReport(housekeeperOid, statementDate);
-
-            if (!report.HasData)
-                return string.Empty;
-
-            report.CreateDocument();
-
-            var filename = Path.Combine(
-                Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments),
-                string.Format("Sandpiper Statement {0:yyyy-MM} {1}.pdf", statementDate, housekeeperName));
-
-            report.ExportToPdf(filename);
-
-            return filename;
-        }
-
-        private static void EmailFile(string emailAddress, string emailBody, string filename, string subject)
-        {
-            var client = new SmtpClient(SystemSettingsHelper.EmailSmtpHost)
-            {
-                Port = SystemSettingsHelper.EmailPort,
-                Credentials =
-                    new NetworkCredential(
-                        SystemSettingsHelper.EmailUsername,
-                        SystemSettingsHelper.EmailPassword)
-            };
-
-            var from = new MailAddress(SystemSettingsHelper.EmailFromEmail, SystemSettingsHelper.EmailFromName,
-                Encoding.UTF8);
-            var to = new MailAddress(emailAddress);
-
-            var message = new MailMessage(from, to)
-            {
-                Subject = subject,
-                SubjectEncoding = Encoding.UTF8,
-                Body = emailBody,
-                BodyEncoding = Encoding.UTF8
-            };
-
-            message.Attachments.Add(new Attachment(filename));
-            client.Send(message);
-            message.Dispose();
-
-            File.Delete(filename);
-        }
-    }
-}
