--- TestChallengeProjects/HouseKeeperHelperProject/XtraMessageBox.cs
@@ -1,9 +1,0 @@
-﻿namespace TestChallengeProjects
-{
-    public class XtraMessageBox
-    {
-        public static void Show(string s, string housekeeperStatements, MessageBoxButtons ok)
-        {
-        }
-    }
-}
