+++ TestChallengeProjects/HouseKeeperProject/Enums/DialogResult.cs
