+++ TestChallengeProjects/HouseKeeperProject/HouseKeeperHelper.cs
@@ -1,0 +1,52 @@
+﻿using TestChallengeProjects.HouseKeeperHelperProject;
+
+namespace TestChallengeProjects
+{
+    public class HouseKeeperHelper
+    {
+        private readonly IUnitOfWork _unitOfWork;
+        private readonly IStatementGenerator _statementGenerator;
+        private readonly IEmailSender _emailSender;
+        private readonly IXtraMessageBox _xtraMessageBox;
+
+        public HouseKeeperHelper(IUnitOfWork unitOfWork, IStatementGenerator statementGenerator, IEmailSender emailSender, IXtraMessageBox xtraMessageBox)
+        {
+            _unitOfWork = unitOfWork;
+            _statementGenerator = statementGenerator;
+            _emailSender = emailSender;
+            _xtraMessageBox = xtraMessageBox;
+        }
+
+        public bool SendStatementEmails(DateTime statementDate)
+        {
+            var housekeepers = _unitOfWork.Query<Housekeeper>();
+
+            foreach (var housekeeper in housekeepers)
+            {
+                if (housekeeper.Email == null)
+                    return false;
+
+                var statementFilename = _statementGenerator.SaveStatement(housekeeper.Oid, housekeeper.FullName, statementDate);
+
+                if (string.IsNullOrWhiteSpace(statementFilename))
+                    return false;
+
+                var emailAddress = housekeeper.Email;
+                var emailBody = housekeeper.StatementEmailBody;
+
+                try
+                {
+                    _emailSender.EmailFile(emailAddress, emailBody, statementFilename,
+                        string.Format("Sandpiper Statement {0:yyyy-MM} {1}", statementDate, housekeeper.FullName));
+                }
+                catch (Exception e)
+                {
+                    _xtraMessageBox.Show(e.Message, string.Format("Email failure: {0}", emailAddress),
+                        MessageBoxButtons.OK);
+                    return false;
+                }
+            }
+            return true;
+        }
+    }
+}
