+++ TestChallengeProjects/HouseKeeperProject/Interfaces/IEmailSender.cs
@@ -1,0 +1,7 @@
+﻿namespace TestChallengeProjects.HouseKeeperHelperProject
+{
+    public interface IEmailSender
+    {
+        void EmailFile(string emailAddress, string emailBody, string filename, string subject);
+    }
+}
