+++ TestChallengeProjects/HouseKeeperProject/Interfaces/IStatementGenerator.cs
@@ -1,0 +1,8 @@
+﻿
+namespace TestChallengeProjects.HouseKeeperHelperProject
+{
+    public interface IStatementGenerator
+    {
+        string SaveStatement(int housekeeperOid, string housekeeperName, DateTime statementDate);
+    }
+}
