+++ TestChallengeProjects/HouseKeeperProject/Interfaces/IXtraMessageBox.cs
@@ -1,0 +1,7 @@
+﻿namespace TestChallengeProjects
+{
+    public interface IXtraMessageBox
+    {
+        void Show(string s, string housekeeperStatements, MessageBoxButtons ok);
+    }
+}
