+++ TestChallengeProjects/HouseKeeperProject/XtraMessageBox.cs
@@ -1,0 +1,9 @@
+﻿namespace TestChallengeProjects
+{
+    public class XtraMessageBox : IXtraMessageBox
+    {
+        public void Show(string s, string housekeeperStatements, MessageBoxButtons ok)
+        {
+        }
+    }
+}
